refactor: organize different reports
Move each report class into its own module.
