+++ memote/suite/reporting/reports/report.py
@@ -15,6 +15,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Subpackage for gathering information and creating reports."""
+"""Render results into pretty HTML."""
 
 from __future__ import absolute_import
+
+from jinja2 import Environment, PackageLoader, select_autoescape
+
+
+class Report(object):
+    """Render a model report."""
+
+    def __init__(self, **kwargs):
+        """Initialize the Jinja2 environment."""
+        super(Report, self).__init__(**kwargs)
+        self.env = Environment(
+            loader=PackageLoader("memote.suite.reporting", "templates"),
+            autoescape=select_autoescape(["html", "xml"])
+        )
+
+    def render_html(self):
+        """Render an HTML report for a model."""
+        raise NotImplementedError("Semi-abstract base class.")
